VIITE-499 Fix DAO so it returns all rows when requested, cleanup.
